persist explorer / endpoint / headers collapsed state in ApiExplorer
