+++ console/src/components/Common/TableCommon/TableStyles.scss
@@ -47,7 +47,6 @@
 
 .table {
   width: -webkit-fill-available;
-  font-size: 14px;
   // width: auto;
 
   thead {
