+++ console/src/components/Services/Data/TableModify/ModifyTable.scss
@@ -155,7 +155,6 @@ hr {
 
 .permissionsTable {
   width: 85%;
-  font-size: 14px;
 
   td, th {
     text-align: center;
